requirements.txt
openwakeword
openai-whisper
pyaudio
numpy
onnxruntime-directml
torch
python-dotenv
